Correct the sizes of devaddrs, net IDs, and TCP ports

+++ priv/repo/migrations/20221007190903_create_routes.exs
@@ -3,7 +3,7 @@ defmodule StartOver.Repo.Migrations.CreateRoutes do
 
   def change do
     create table("organizations", primary_key: false) do
-      add :oui, :integer, primary_key: true
+      add :oui, :bigint, primary_key: true
       add :owner_wallet_id, :string, null: false
       add :payer_wallet_id, :string, null: false
 
@@ -12,7 +12,7 @@ defmodule StartOver.Repo.Migrations.CreateRoutes do
 
     create table("routes") do
       add :oui, references(:organizations, column: :oui, on_delete: :delete_all)
-      add :net_id, :integer, null: false
+      add :net_id, :bigint, null: false
 
       timestamps()
     end
@@ -36,7 +36,7 @@ defmodule StartOver.Repo.Migrations.CreateRoutes do
     create table("route_servers") do
       add :route_id, references(:routes, on_delete: :delete_all), primary_key: true
       add :host, :string, null: false
-      add :port, :integer, null: false
+      add :port, :bigint, null: false
       add :protocol_opts, :map, null: false
 
       timestamps()
